Csfr

+++ to-do-app/backend/ToDoApp.API/ToDoApp.API/Controllers/TasksController.cs
@@ -1,4 +1,6 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using System.Security.Claims;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using ToDoApp.API.Models.Domain;
 using ToDoApp.API.Models.Dtos;
 using ToDoApp.API.Repositories.Interfaces;
@@ -6,6 +8,7 @@ using ToDoApp.API.Repositories.Interfaces;
 namespace ToDoApp.API.Controllers;
 
 [ApiController]
+[Authorize]
 [Route("api/[controller]")]
 public class TasksController : ControllerBase
 {
@@ -20,24 +23,25 @@ public class TasksController : ControllerBase
     [Route("{taskId:Guid}")]
     public async Task<IActionResult> GetByIdAsync([FromRoute] Guid taskId)
     {
+        var userId = User.FindFirstValue("sub");
+        if (string.IsNullOrEmpty(userId)) return Unauthorized();
+
         var task = await _tasksRepository.GetByIdAsync(taskId);
-        if (task == null)
-        {
-            return NotFound();
-        }
+        if (task == null) return NotFound();
+
+        if (task.OwnerId != Guid.Parse(userId)) return Forbid();
 
         return Ok(task);
     }
 
     [HttpGet]
+    [Authorize(Roles = "Admin")]
     public async Task<IActionResult> GetAllAsync()
     {
-        var tasks = await _tasksRepository.GetAllAsync();
-        if (tasks == null)
-        {
-            return BadRequest();
-        }
+        var userId = User.FindFirstValue("sub");
+        if (string.IsNullOrEmpty(userId)) return Unauthorized();
 
+        var tasks = await _tasksRepository.GetAllByOwnerIdAsync(Guid.Parse(userId));
         return Ok(tasks);
     }
 
@@ -45,22 +49,25 @@ public class TasksController : ControllerBase
     [Route("owner/{ownerId:Guid}")]
     public async Task<IActionResult> GetAllByOwnerIdAsync([FromRoute] Guid ownerId)
     {
-        var tasks = await _tasksRepository.GetAllByOwnerIdAsync(ownerId);
-        if (tasks == null)
-        {
-            return BadRequest();
-        }
+        var userId = User.FindFirstValue("sub");
+        if (string.IsNullOrEmpty(userId)) return Unauthorized();
 
+        var tasks = await _tasksRepository.GetAllByOwnerIdAsync(Guid.Parse(userId));
         return Ok(tasks);
     }
 
     [HttpPost]
     public async Task<IActionResult> CreateAsync([FromBody] CreateTaskRequest request)
     {
+        var userId = User.FindFirstValue("sub");
+        if (string.IsNullOrEmpty(userId))
+        {
+            return Unauthorized();
+        }
         var task = new TaskItem
         {
             Text = request.Text,
-            OwnerId = request.OwnerId
+            OwnerId = Guid.Parse(userId)
         };
 
         var createdTaskId = await _tasksRepository.CreateAsync(task);
@@ -72,12 +79,16 @@ public class TasksController : ControllerBase
     [Route("{taskId:Guid}")]
     public async Task<IActionResult> DeleteAsync([FromRoute] Guid taskId)
     {
-        var isDeleted = await _tasksRepository.DeleteByIdAsync(taskId);
-        if (!isDeleted)
-        {
-            return NotFound();
-        }
+        var userId = User.FindFirstValue("sub");
+        if (userId == null) return Unauthorized();
 
-        return NoContent();
+        var task = await _tasksRepository.GetByIdAsync(taskId);
+        if (task == null) return NotFound();
+
+        if (task.OwnerId != Guid.Parse(userId))
+            return Forbid();
+
+        var isDeleted = await _tasksRepository.DeleteByIdAsync(taskId);
+        return isDeleted ? NoContent() : BadRequest();
     }
 }
